+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Improved compiler error message when using `AnyScan` with incompatible pairs due
   to `first_type` being const on the right side.
 * Added concept `IsSameAsAnyOfRaw` / `NotSameAsAnyOfRaw` which determine whether type is one of a list of types.
+* Added `const char* tstring::c_str` which is occasionally needed.
 
 # 0.2.12